Move "IHasPrimaryKey" into more general namespace.
And note that there's no need to place the `Beatmaps/Types` namespace because only type in the `Objects` did that.
