+++ osu.Game.Rulesets.Karaoke/Beatmaps/Stages/StageElementCategory.cs
@@ -9,7 +9,6 @@ using System.Linq;
 using Newtonsoft.Json;
 using osu.Framework.Bindables;
 using osu.Game.Rulesets.Karaoke.Objects;
-using osu.Game.Rulesets.Karaoke.Objects.Types;
 
 namespace osu.Game.Rulesets.Karaoke.Beatmaps.Stages;
